fdf 1.8

+++ fdf/include/fdf.h
@@ -6,7 +6,7 @@
 /*   By: gdoumer <gdoumer@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/01/18 15:55:30 by gdoumer           #+#    #+#             */
-/*   Updated: 2024/02/04 00:15:55 by gdoumer          ###   ########.fr       */
+/*   Updated: 2024/02/04 00:39:03 by gdoumer          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -37,7 +37,7 @@ typedef struct s_struc_array
 	int			len_raw;
 	int			start_x;
 	int			start_y;
-	int			ratio;
+	float		ratio;
 	int			height_of_z;
 }	t_stray;
